cloud/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
open3d==0.18.0
paramiko==3.3.1
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
google-generativeai==0.3.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
